+++ CHANGELOG.md
@@ -3,6 +3,12 @@ All notable changes to this project will be documented in this file.
 
 The format is based on [Keep a Changelog](http://keepachangelog.com/).
 
+## [Unreleased]
+
+### Added
+
+- Support for using OKE Workload Identity to authenticate to OCI
+
 ## 1.0.18 - July 2024
 
 ### Added